requirements.txt
asgiref==3.6.0
crispy-bootstrap4==2022.1
Django==4.0
django-crispy-forms==2.0
django-phonenumber-field==7.1.0
phonenumbers==8.13.11
Pillow==9.5.0
psycopg2==2.9.6
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
